refactor(test/e2e): split default values and personas test-completion specs
